requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

annotated-types==0.7.0
anyio==4.6.2.post1
atproto==0.0.55
azure-core==1.32.0
azure-functions==1.21.3
azure-storage-blob==12.24.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
dill==0.3.8
dnspython==2.7.0
feedparser==6.0.11
fonttools==4.53.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inquirerpy==0.3.4
isodate==0.7.2
joblib==1.4.2
kiwisolver==1.4.5
kneed==0.8.5
libipld==3.0.0
lightgbm==4.5.0
#matplotlib==3.9.1
multiprocess==0.70.14
nestd==0.3.1
numpy==2.0.0
orjson==3.9.15
packaging==24.1
pandas==2.2.2
patsy==0.5.6
pfzy==0.3.4
pillow==10.4.0
prompt_toolkit==3.0.47
PuLP==2.9.0
pycparser==2.22
pydantic==2.10.2
pydantic_core==2.27.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
robyn==0.56.2
rustimport==1.3.4
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.1
statsmodels==0.14.2
threadpoolctl==3.5.0
toml==0.10.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
watchdog==2.2.1
wcwidth==0.2.13
websockets==13.1